merge dev to main for last minute review

+++ app/models/transaction.py
@@ -13,7 +13,7 @@ class Transaction(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod('users.id')), nullable=False)
     company_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod(
         'companies.id')), nullable=False)
-    total = db.Column(db.Float(2), nullable=False)
+    total = db.Column(db.Float(6), nullable=False)
     type = db.Column(db.String, nullable=False)
     created_at = db.Column(db.DateTime, default=datetime.now, nullable=False)
     updated_at = db.Column(db.DateTime, default=datetime.now,
